refactor: restructure tests to be feature-conditional
Redesign integration tests to ensure cargo test works without feature flags:

**Changes:**
- Move feature-specific tests from integration tests into module tests
- Create basic_integration.rs with only default-feature tests
- Move JSON tests to utils module with #[cfg(feature = "json")]
- Move progress tracking and async tests to mod.rs with proper feature gates
- Move S3 upload test to s3.rs module with early validation test

**Benefits:**
- cargo test now passes without any feature flags
- All tests are properly feature-gated and skip gracefully when disabled
- Tests are co-located with the code they test
- No compilation failures on minimal feature sets
- Better test organization by functionality

**Test structure:**
- tests/basic_integration.rs: Default features only (gz, bz, http)
- src/oneio/mod.rs: Progress tracking and async tests with feature gates
- src/oneio/utils.rs: JSON functionality tests
- src/oneio/s3.rs: S3-specific tests including upload validation

Resolves cargo test failures on default feature configuration.
