+++ Cargo.toml
@@ -91,7 +91,7 @@ cli = ["clap", "tracing", "gz", "bz", "http", "s3", "digest"]
 # Advanced: TLS selection (only if explicitly needed)
 native-tls = [
     "reqwest?/native-tls",
-    "suppaftp?/native-tls", 
+    "suppaftp?/native-tls",
     "rust-s3?/sync-native-tls",
 ]
 rustls = [
