+++ src/oneio/utils.rs
@@ -90,3 +90,30 @@ pub fn read_lines(path: &str) -> Result<Lines<BufReader<Box<dyn Read + Send>>>, 
     let buf_reader = BufReader::new(reader);
     Ok(buf_reader.lines())
 }
+
+#[cfg(test)]
+mod tests {
+
+    #[cfg(feature = "json")]
+    #[test]
+    fn test_read_json_struct() {
+        #[derive(serde::Deserialize, Debug)]
+        struct Data {
+            purpose: String,
+            version: u32,
+            meta: SubData,
+        }
+        #[derive(serde::Deserialize, Debug)]
+        struct SubData {
+            float: f64,
+            success: bool,
+        }
+
+        let data = crate::read_json_struct::<Data>("https://spaces.bgpkit.org/oneio/test_data.json").unwrap();
+
+        assert_eq!(data.purpose, "test".to_string());
+        assert_eq!(data.version, 1);
+        assert_eq!(data.meta.float, 1.1);
+        assert_eq!(data.meta.success, true);
+    }
+}
